requirements.txt
# Cold Email Automation App Dependencies
# Core framework
streamlit>=1.28.0

# Google APIs
google-api-python-client==2.100.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0

# AI Integration
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.3
numpy>=1.24.0
openpyxl==3.1.2

# Environment management
python-dotenv==1.0.0

# Additional utilities
requests==2.31.0
PyPDF2==3.0.1
